requirements.txt
gunicorn>=21.2.0
uvicorn>=0.23.0
fastapi>=0.100.0
fastf1>=3.3.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.31.0
httpx>=0.24.0
google-genai>=1.0.0
pydantic>=2.0.0
